fix sql handler column type to decimal

+++ indexer/packages/postgres/src/db/migrations/migration_files/20250514151100_add_builder_fees_to_orders_and_fills.ts
@@ -8,7 +8,7 @@ export async function up(knex: Knex): Promise<void> {
 
   await knex.schema.alterTable('fills', (table) => {
     table.string('builderAddress').nullable().defaultTo(null);
-    table.bigInteger('builderFee').nullable().defaultTo(null);
+    table.decimal('builderFee').nullable().defaultTo(null);
   });
 }
